Enhance calibration and error handling in flight controller and sensor libraries
- Added error handling for BMP280 initialization in MioBMP280.cpp
- Improved offset calculation in MioQMC5883L.cpp for better accuracy
- Updated flight controller sketch to include delta time in control steps
- Refactored joystick control logic in sketch_controller.ino for better calibration

+++ libreries/Mio_QMC5883L/MioQMC5883L.cpp
@@ -152,9 +152,9 @@ bool MioQMC5883L::calibrate(uint32_t durationMs) {
     }
 
     // Calcola offset raw
-    _offsetRawX = (minX + maxX) / 2;
-    _offsetRawY = (minY + maxY) / 2;
-    _offsetRawZ = (minZ + maxZ) / 2;
+    _offsetRawX = (int16_t)(((int32_t)minX + (int32_t)maxX) / 2);
+    _offsetRawY = (int16_t)(((int32_t)minY + (int32_t)maxY) / 2);
+    _offsetRawZ = (int16_t)(((int32_t)minZ + (int32_t)maxZ) / 2);
 
     // Calcola offset in Gauss
     float scale = getScale();
